Compatibility with DataTables 1.13's ES modules
Now that DataTables provides ES modules, Vite will automatically use them rather than the UMD files. However, the signature of the ESM files is slightly different from the CommonJS loader that was used before. We no longer need to execute a function - it is just present automatically from the import.

I've also shortened the setting of how the default properties are set. It does exactly the same thing.